:art: Logicを別componentに移してテストを書いた
