+++ src/components/container/useViewModel.ts
@@ -2,6 +2,7 @@ import { useContext, useState } from "react"
 
 import { usePopulationComposition } from "@/components/hooks/api/usePopulationComposition"
 import { usePrefectures } from "@/components/hooks/api/usePrefectures"
+import { LoadingContext } from "@/components/hooks/useLoading"
 import {
   ALL_POPULATION,
   POPULATION_BY_OLDER,
@@ -9,7 +10,6 @@ import {
   POPULATION_BY_YOUNG,
   PopulationCompositionType,
 } from "@/constants/populationCompositionType"
-import { LoadingContext } from "@/pages"
 import { CustomHook } from "@/types/customHook"
 import {
   PopulationCompositionGraphElements,
