+++ src/components/hooks/useLoading.tsx
@@ -1,0 +1,15 @@
+import { createContext, useState } from "react"
+
+type LoadingContextType = {
+  isLoading: boolean
+  setIsLoading: (isLoading: boolean) => void
+}
+
+export const LoadingContext = createContext<LoadingContextType>({
+  isLoading: false,
+  setIsLoading: () => {},
+})
+export const useLoading = () => {
+  const [isLoading, setIsLoading] = useState(false)
+  return { isLoading, setIsLoading }
+}
